ENH: add probability class prediction and ROC calculation

+++ anoII/parkinglot/classifier/classifier_set.h
@@ -11,9 +11,9 @@ class ClassifierSet
 {
 public:
 	void train(const std::vector<Example>& examples);
-	std::vector<int> predict(const Example& example, cv::Mat frame);
-	std::vector<std::vector<int>> predictMultiple(const std::vector<std::unique_ptr<Extractor>>& extractors, const std::vector<cv::Mat>& frames);
-	int predictClass(const std::vector<int>& response);
+	std::vector<float> predict(const Example& example, cv::Mat frame);
+	std::vector<std::vector<float>> predictMultiple(const std::vector<std::unique_ptr<Extractor>>& extractors, const std::vector<cv::Mat>& frames);
+	int predictClass(const std::vector<float>& response);
 
 	void load(const std::string& name);
 	void save(const std::string& name);
